Move viewport step to trait
